refactor: move QS functions to querystring.ts
