+++ src/ts/view/Home.ts
@@ -6,12 +6,14 @@ import { IProducts } from '../types/interface';
 import { ProductCard } from '../service/StoreService';
 import { search } from '../helpers/search';
 import { generateFilter, recalcFilters, filter } from '../helpers/filters';
-
+import { restoreFiltersFromQS, saveFiltersToQS } from '../helpers/querystring';
 import { Cart } from '../service/Cart';
+import { cartSum } from '../helpers/addProduct';
+
 const shopCart = new Cart();
 const cartCounter = document.querySelector('.cart-counter');
 const cartTotal = document.querySelector('.cart-total__price');
-import { cartSum } from '../helpers/addProduct';
+
 export class Home extends AbstractView {
     constructor(params: QueryStringParams) {
         super(params);
@@ -105,16 +107,6 @@ export class Home extends AbstractView {
     `;
     }
 
-    updateQueryStringParameter(uri: string, key: string, value: string) {
-        var re = new RegExp('([?&])' + key + '=.*?(&|$)', 'i');
-        var separator = uri.indexOf('?') !== -1 ? '&' : '?';
-        if (uri.match(re)) {
-            return uri.replace(re, '$1' + key + '=' + value + '$2');
-        } else {
-            return uri + separator + key + '=' + value;
-        }
-    }
-
     convertToProductCard(products: IProducts): ProductCard[] {
         let productCards: ProductCard[] = [];
         for (let item in products) {
@@ -140,42 +132,15 @@ export class Home extends AbstractView {
     }
 
     async prepareSearch(caller: string) {
-        let filterSearch: string = (document.querySelector('.product-search') as HTMLInputElement).value.toLowerCase();
-        let filterSort: string = (document.querySelector('.product-sort') as HTMLSelectElement).value;
         const products = (await loadProducts()) as unknown as IProducts;
         let productCards = this.convertToProductCard(products);
         productCards = filter(productCards);
+        let filterSearch: string = (document.querySelector('.product-search') as HTMLInputElement).value.toLowerCase();
+        let filterSort: string = (document.querySelector('.product-sort') as HTMLSelectElement).value;
         productCards = search(productCards, filterSearch, filterSort);
         recalcFilters(productCards, caller);
         this.drawProductCards(productCards);
-
-        let url: string = window.location.href;
-
-        let filterCategory = Array.from(document.querySelectorAll('.filter-category-list .filter-check') as unknown as HTMLInputElement[]);
-        filterCategory = filterCategory.filter((i) => i.checked === true);
-        let filterCategoryValues = filterCategory.map((i) => i.id).join('|');
-        url = this.updateQueryStringParameter(url, 'category', filterCategoryValues);
-
-        let filterBrand = Array.from(document.querySelectorAll('.filter-brand-list .filter-check') as unknown as HTMLInputElement[]);
-        filterBrand = filterBrand.filter((i) => i.checked === true);
-        let filterBrandValues = filterBrand.map((i) => i.id).join('|');
-        url = this.updateQueryStringParameter(url, 'brand', filterBrandValues);
-
-        const priceInput = document.querySelectorAll(`.price-range .price-input input`) as unknown as HTMLInputElement[];
-        let minprice = priceInput[0].value;
-        let maxprice = priceInput[1].value;
-        url = this.updateQueryStringParameter(url, 'minprice', minprice);
-        url = this.updateQueryStringParameter(url, 'maxprice', maxprice);
-
-        const stockInput = document.querySelectorAll(`.stock-range .price-input input`) as unknown as HTMLInputElement[];
-        let minstock = stockInput[0].value;
-        let maxstock = stockInput[1].value;
-        url = this.updateQueryStringParameter(url, 'minstock', minstock);
-        url = this.updateQueryStringParameter(url, 'maxstock', maxstock);
-
-        url = this.updateQueryStringParameter(url, 'search', filterSearch);
-        if (filterSort !== '') url = this.updateQueryStringParameter(url, 'sort', filterSort);
-        window.history.pushState({ path: url }, '', url);
+        saveFiltersToQS();
     }
 
     sliderListener(slider: string) {
@@ -230,77 +195,12 @@ export class Home extends AbstractView {
         let productCards = this.convertToProductCard(products);
         generateFilter(productCards, 'category');
         generateFilter(productCards, 'brand');
-
         let params = new URL(window.location.href).searchParams;
-
-        //restore 'search' from QS
-        let searchParam = params.has('search') ? params.get('search') : '';
-        (document.querySelector('.product-search') as HTMLInputElement).value = searchParam as string;
-
-        //restore 'sort' from QS
-        let sortParam = params.has('sort') ? params.get('sort') : '';
-        (document.querySelector('.product-sort') as HTMLSelectElement).value = sortParam as string;
-
-        //restore 'category' from QS
-        let categoryParam = params.has('category') ? params.get('category') : '';
-        if (categoryParam !== '') {
-            let filterCategory = Array.from(
-                document.querySelectorAll('.filter-category-list .filter-check') as unknown as HTMLInputElement[]
-            );
-            let filterCategoryValues = categoryParam!.split('|');
-            filterCategory.forEach(element => {
-                if (filterCategoryValues.includes(element.id)) element.checked = true;
-            });
-        }
-
-        //restore 'brand' from QS
-        let brandParam = params.has('brand') ? params.get('brand') : '';
-        if (brandParam !== '') {
-            let filterBrand = Array.from(
-                document.querySelectorAll('.filter-brand-list .filter-check') as unknown as HTMLInputElement[]
-            );
-            let filterBrandValues = brandParam!.split('|');
-            filterBrand.forEach(element => {
-                if (filterBrandValues.includes(element.id)) element.checked = true;
-            });
-        }
-
-        //restore min & max 'price' from QS
-        const minpriceParam = params.has('minprice') ? params.get('minprice') : '';
-        const maxpriceParam = params.has('maxprice') ? params.get('maxprice') : '';
-        let priceInput = document.querySelectorAll(`.price-range .price-input input`) as unknown as HTMLInputElement[];
-        let rangeInput = document.querySelectorAll(`.price-range .range-input input`) as unknown as HTMLInputElement[];
-        let range = document.querySelector(`.price-range .slider .progress`) as HTMLDivElement;
-        if (minpriceParam !== '') {
-            priceInput[0].value = minpriceParam!.toString();
-            rangeInput[0].value = minpriceParam!.toString();
-            range!.style.left = (Number(minpriceParam!) / Number(rangeInput[0].max)) * 100 + '%';
-        }
-        if (maxpriceParam !== '') {
-            priceInput[1].value = maxpriceParam!.toString();
-            rangeInput[1].value = maxpriceParam!.toString();
-            range!.style.right = 100 - (Number(maxpriceParam) / Number(rangeInput[1].max)) * 100 + '%';
-        }
-
-        //restore min & max 'in stock' from QS
-        const minstockParam = params.has('minstock') ? params.get('minstock') : '';
-        const maxstockParam = params.has('maxstock') ? params.get('maxstock') : '';
-        priceInput = document.querySelectorAll(`.stock-range .price-input input`) as unknown as HTMLInputElement[];
-        rangeInput = document.querySelectorAll(`.stock-range .range-input input`) as unknown as HTMLInputElement[];
-        range = document.querySelector(`.stock-range .slider .progress`) as HTMLDivElement;
-        if (minstockParam !== '') {
-            priceInput[0].value = minstockParam!.toString();
-            rangeInput[0].value = minstockParam!.toString();
-            range!.style.left = (Number(minstockParam!) / Number(rangeInput[0].max)) * 100 + '%';
-        }
-        if (maxstockParam !== '') {
-            priceInput[1].value = maxstockParam!.toString();
-            rangeInput[1].value = maxstockParam!.toString();
-            range!.style.right = 100 - (Number(maxstockParam) / Number(rangeInput[1].max)) * 100 + '%';
-        }
-
+        restoreFiltersFromQS(params);
         productCards = filter(productCards);
-        productCards = search(productCards, searchParam!, sortParam!);
+        let filterSearch: string = (document.querySelector('.product-search') as HTMLInputElement).value.toLowerCase();
+        let filterSort: string = (document.querySelector('.product-sort') as HTMLSelectElement).value;
+        productCards = search(productCards, filterSearch, filterSort);
         recalcFilters(productCards, 'firstload');
         this.drawProductCards(productCards);
         this.bindListeners();
